Extract `collection/snapshots.rs`
